Create MM20230107_analysis_CFU_count_boxplot.Rmd
Analysis of CFU counts. The script first analyses and plots the data without normalisation. Afterwards, the data are normalised and the analysis/ plotting are repeated. It is best to compare the output of both analyises and decide if normalisation is required. Also, check if ANOVA across all batches or Dunnet's test comparing each sample to a control per experiment are better suitable for statistical analysis.